Convert GameState to TypeScript
